requirements.txt
braintree==3.49.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.9.1
requests==2.20.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
